Fixes #1034
The eventstore made the wrong conversion and returned a negative value
for totalbytessent in JSON response from the
http://*:1114/stats/replication API call.

+++ src/EventStore.Core/Services/Replication/MasterReplicationService.cs
@@ -161,7 +161,7 @@ namespace EventStore.Core.Services.Replication
                     if (subscription.Value != null)
                     {
                         var stats = new ReplicationMessage.ReplicationStats(subscription.Key, tcpConn.ConnectionId, subscription.Value.ReplicaEndPoint.ToString(), tcpConn.SendQueueSize,
-                                            (int)conn.TotalBytesSent, (int)conn.TotalBytesReceived, conn.PendingSendBytes, conn.PendingReceivedBytes);
+                                            conn.TotalBytesSent, conn.TotalBytesReceived, conn.PendingSendBytes, conn.PendingReceivedBytes);
                         replicaStats.Add(stats);
                     }
                 }
